all: force use of vm.Malloc
Add a check in RowSplitter that every call to Write
passes a buffer from vm.Malloc, and then fix all
the sources of input buffers to the query engine.

(Some of these are pessimized because of copying,
but it should be straightforward enough to fix them.
There are also a few places where we generate errors
if the intermediate results of operations lead to more
than 1MiB of output.)