vm: perform just-in-type copying in Splitter
Relax the core constraints a bit so that copying
happens in vm.RowSplitter.Write. It is in the best
place to determine how to split up the input data
so that it always fits in one "page," and it can
choose to copy a small amount of data at a time
if it wants. (We still want the normal case to
have data copied in VMM.)