vm: add -vmfence option
If the output binary is built with -tags=vmfence,
then we will perform a bunch of additional (expensive)
memory protection operations on the VMM region to
guard against out-of-bounds reads and writes.

Since ~all memory accesses in the VM are now bounded
to the VMM region, this should catch most bad accesses.