exprcompile: implement "generic case"
Since we can blend together the results
of arbitrary expressions via some combination
of re-boxing and regular value parsing,
we can eliminate the special case for serialized
results and just fall back to compiling everything
that way unless we have a more precise hint about
the expected type of the CASE expression.

Also, fix a bug in which falseprop was eliminating
blend{v,s} instructions; these instructions still have
a semantic meaning when the mask argument is false,
so the need to be simplified slightly differently.