ion/blockfmt: set max decoders to GOMAXPROCS
Noticed that parallel decoding seemed slow,
and a block profile revealed that compress/zstd
was only allocating buffers for 4 simultaneous decoders.
(The docs describe this behavior, but I hadn't read
them carefully enough.)

Set the maximum parallel decoder to GOMAXPROCS
so that calls to zstd.Decoder.DecodeAll do not block
on high-core-count machines.