vm: make non-vmm writes more flexible
Make the handling of non-vmm memory
passed to Write more robust:

Update scan() so that it accepts
buffers that have been cut off prematurely
(via slicing or copying too few bytes).
This means we can limit the number of bytes
that are copied in a single pass inside
RowSplitter.Write by trying to copy just
a few kB at a time and splitting it and
seeing if we get enough delimiters.
(The only thing we can't handle are individual
objects larger than 1 page, but we have
that limitation elsewhere as well.)

Also, update a few places where we were
explicitly allocating vmm buffers to
*not* do that. In many cases those places
are actually writing directly to output,
so allocating a buffer and copying is pure
overhead. Similarly, if a copy *is* necessary,
it is better that we centralize the logic
in RowSplitter so that we can control it.