vm: remove explicit buffers; add vmref
For all [][2]uint32 that are actually
supposed to be addresses relative to vmm,
convert those to []vmref and remove the
explicit base pointer passed with those
functions.

The few cases left where we use [][2]uint32
are in testing, where we are calling functions
that we know do not require vmm-relative
addressing.