vm: fix scanning of nop pads
In the unlikely even that scanvmm() sees
a nop pad we may get back more than 0 bytes
to skip even when the number of delimiters
produced is zero. That's an acceptable result;
we should not panic in that case.

Add some coverage in the test code that
confirms we eat standalone nop pads correctly.