vm: try harder to malloc
Track the number of available pages explicitly
and only fail an allocation if we are actually
out of pages. The bits in the bitmap may be
temporarily flipped while pages are being free'd
by the kernel; those pages aren't "in use" and
will likely become usable again shortly.

This prevents the race where many bitmaps are
being madvise(2)'d simultaneously, which *looks*
like an out-of-memory condition, but really isn't.