cmd/snellerd: return better errors
If we get anything that looks like a
compilation error, that should be returned
back to the user rather than the mostly-useless
"cannot create query plan" message.

Similarly, if we encounter a parse error,
return that directly as well.

Add pir.CompileError, which will include
contextual information on compilation errors
when they are present.