db: implement precise GC
Rather than performing a list operation
to determine the list of stale objects,
track stale objects explicitly in the index
and remove expired objects each time the
index is re-loaded for ingest.

Listing for stale objects has the advantage
that it will catch objects that were partially-written
(i.e. due to a crash/outage, etc.), but it has the
drawback that it runs in O(N) time with respect
to the total number of input objects in the table.
For extremely large tables, this will make listing
too slow to run inside our 3-second latency target.

(This will be particularly important for reducing
the GC overhead for B-tree index contents, which
is coming shortly; we don't want to have to walk
the B-tree to determine the list of live packed-*.ion.zst
files!)

I have left the "traditional" GC enabled for now
so that we can see if precise GC works well alongside it,
but once we have some confidence in precise GC we
can remove the traditional GC entirely.