blockfmt, db: implement indirect index references
Replace Index.Contents -> Index.Inline and add
Index.Indirect, which is an ordered list of indirect
blocks containing a compressed list of descriptors.

When the total decompressed size of Index.Inline
exceeds some threshold, we flush those descriptors
to an indirect list and just stick a summary of the
block metadata in the index. (Right now that threshold
is 100GiB, which means ~1000x 100MiB decompressed spans,
so we'd keep a maximum of 1000 block desriptors in the index.)

plan.Env is adjusted so that Stat() can produce a TableHandle
that incorporates predicate information, and db.Blobs is similarly
adjusted so that we only produce blobs from descriptors in the
index that actually match the predicate. In practice this means
we can usually avoid loading the indirect descriptors (assuming
most queries want to touch data that was ingested most recently).