all: simplify plan.TableHandle / plan.Env
The plan.Env API was a bit odd and cumbersome
because there were two "return values" for
the Stat operation: one in expr.Table.Opaque,
and the other as plan.TableHandle.
(I'm the who wrote this code originally and
it confuses *me*, which is not a great sign.)

Instead, eliminate expr.Table.Opaque and put
all the table handle responsibilities into
plan.TableHandle, as it should be.
TableHandles now need to know how to serialize
themselves as ion data, and callers of plan.Decode
(i.e. tnproto.Serve, etc.) need to know how to
deserialize the TableHandles that they expect to receive.
(In practice this is just snellerd's *main.filterTable).

So, plan.Env only needs to exist during planning,
and not as part of execution, (de)serialization, etc,
which makes the separation of responsibilities a bit
clearer IMO.

Also, there are some improvements in plan/ as far
as how we go about Stat()ing table elements.
We now guarantee that we Stat() each table reference
exactly once (and *not* during deserialization), taking
care to re-use the handle from a previous Stat() call
when calling plan.Splitter.Split().

There is probably more that could be done to make
this a bit more boilerplate-free, but I think this
is a good first step.