db: implement sync as scan
There were at least two pretty bad bugs
in db.Builder.Sync due to that code path
rotting in favor of the better (incremental)
APIs we have for adding data.

Instead of maintaining 'sync' as a separate
operation, we can just implement it as a
forced scan. We get a bunch of good stuff
"for free" from re-using the scanning code;
one big improvement is that we get incremental
ingest "for free" on each sync invocation,
and we are guaranteed to list from where we
last left off.