all: restrict ORDER BY size
Bring ORDER BY restrictions in line with
the documentation: we will only perform
and ORDER BY if it is operating on
a) a "column cardinality" result (DISTINCT, GROUP BY),
since those already have output limits, or
b) an explicitly-limited output set.

This restriction is applied after optimization,
so it doesn't apply to cases where the ORDER BY
can be eliminated or where we can infer statically
that the number of rows that can reach the ORDER BY
is known to be sufficiently small.

I've also fixed a bug in multi-column ktop sorting
that was exposed by converting one of the plan/TestExec
cases from an un-limited to a limited query.