expr/blob: intern common trailers
Introduce blob.CompressedPart, which refers
to a subrange of blocks within blob.Compressed.

Each blob.CompressedPart can point to a *shared*
blob.Compressed, and during encoding and decoding
we only serialize the full blob.Compressed parent
once, and subsequently we just refer to it by an
integer ID that is tracked during encoding and decoding.

In high-compression-ratio inputs that have many
blocks per packed file, we can reduce the encoded
size of the *split* list of blobs for that table
by an order of magnitude or more.
(However, partitioning reduces the compression improvement
here, as we encode one blob.List per subtable, and we
don't currently intern each blob.Compressed across all
the different blob.List / subtables. Perhaps that would
be a good follow-up improvement.)