blockfmt: add prefetching to Converter.Run
If we are converting data from e.g. S3, where
the time-to-first-byte for each input stream is
quite large, we can get a pretty substantial increase
in overall throughput by starting requests well in advance
of actually needing the response bytes.
(Without prefetching, our peformance on a large list
of very small input files is not terribly good, because
we spend all of our time waiting for the response bytes
and almost zero time doing any real work.)

Right now I've picked a bandwidth-delay product of 80MB,
but we can tune that later if it turns out to be wildly wrong.

The readahead algorithm for single-stream conversion
is pretty straightforward; we simply perform an asynchronous
zero-byte read for future inputs up to some limit (in bytes
and in number of outstanding reads).
For parallel conversion, there is an intermediate set of
goroutines that perform the prefetching and then write to
the channel that the conversion goroutines are listening on.

In practice this is worth about a 10x perf improvement
(in wall clock time per byte of input) for S3 workloads
on S3 that use many small files.