expr/blob: use snapshot to rewind when compressing (#1041)
When compressing a blob list, we write the list into the buffer, and if
it's too large we rewind the buffer and compress the list. This causes
the segment tracking in the buffer to be reset, which can cause panics
when writing compressed blob lists inside composite values.

This fixes that by using snapshots to safely rewind the buffer.