aws: keep 2 explicit clamped keys
For the upcoming INTO support, we'll need
to be able to serialize aws signing keys,
so we can't keep the result of hmac.New
in aws.SigningKey. Use an explicit []byte
containing the "derived" key instead.

Also, keep around a second derived key for
use when we pass through UTC midnight,
and update one of the unit tests so that
we get some coverage for that scenario.