vm: add sort prefiltering
When we are performing a sorting operation like
  ORDER BY x LIMIT N
we can use the (N-1)'th element to compile some
bytecode that filters out elements larger/smaller
than that element.

Currently we can only do prefiltering for
numeric or timestamp results, and only for ORDER BY clauses
that sort on one column.