internal/sort: re-use record storage
When we are adding a record to the heap,
we can clobber the storage used by the
previous "max" entry rather than allocating
new memory for a copy of the record data.

In cases where we have a medium-sized number
of heap entries or a large number of records
that make it through prefiltering, this ends
up saving many thousands of allocations.