ion: improve Chunker
A couple fixes and impprovements to ion.Chunker:

Allow significantly higher-cardinality inputs
before erroring out by resetting the symbol table
when we flush blocks (this would be at 100MB boundaries,
so this wouldn't occur too frequently in practice).
In order to reset the symbol table, we need to take
care to adjust all of the symbols in the un-flushed
structure(s) still left in the buffer.

This change also revealed to bugs:

 1. (*Chunker).forceFlush was flushing the trailing structure
    when it fit into the alignment, even if the symbol table
    had not been adjusted to the correct position; we can only
    flush the last struct if c.tmpID == c.Symbols.MaxID()
 2. (*Chunker).Write cannot use the c.Symbols.Contains(&newsyms)
    check, because the additional symbols in c.Symbols would not
    be overwritten correctly by subsequent incremental symbol
    tables.