plan: add optional support for env providing index (#1128)
This adds the plan.Indexer interface which allows plan.Env
implementations to optionally provide an index for a table to be used
during plan optimization. This also adds support for combined indexes
for multiple tables when using glob expressions or the append operator,
allowing sparse indexing to be used to optimize such queries.