all: introduce blockfmt.SparseIndex
Use a columnar data structure for storing
time intervals rather than storing min/max
values in each block. This saves a factor
of 5 to 10 in index and query plan size.

The new Trailer.Sparse and IndirectTree.Sparse
are populated during unmarshaling when we
encounter old-format indices, so these data structures
are backwards-compatible with what we already have
on-disk.

The blockfmt.TimeIndex data structure stores
an *interpolation* of the time values that
are inserted into each block. When blocks
have overlapping ranges, those are compressed away,
so indexed values that are very non-monotonic don't
end up consuming much space. Values that are inserted
perfectly monotonically will have the same level of
indexing precision as before, but with less space overhead.
The list of min- and max-time values and offsets
are delta-encoded when serialized, which saves a bunch
of additional space.