ion/blockfmt: various cleanup + improvements
- Remove Trailer.Slice, since it is dangerous
(it doesn't update Sparse)

- Use TrailerDecoder for decoding TimeIndex.{min,max},
which cuts down on total allocations by a factor of two
or three.

- Double-delta-encode the integer sequences for TimeIndex.{min,max},
since the intervals tend to be somewhat equally spaced, and smaller
magnitude integers consume fewer ion-encoded bytes.
(A perfectly-spaced sequence of offsets and timestamps consumes
only two bytes per interval in the limit.)