Ion/blockfmt: delta-encode trailer blocks
For Trailer.Blocks, double-differential-encode the
offsets and delta-encode the chunk counts.
(Offsets tend to be evenly-spaced, and chunk counts
tend to be similar in adjacent blocks, so we can
minimize the encoded integer magnitudes by encoding
their second- and first-derivative values, respectively.)

In the synthetic BenchmarkIndexDecodingAllocs example,
this change ends up cutting the size of the encoded
index in half.