ion, jsonrl: limit sparse index data collection
Don't collect sparse indexing information
for every single timestamp value.
If the value is not present in at least 1/3rd
of the input rows, or the depth of the object
exceeds 3, then we can ignore the indexing info.

(The github archive dataset has ~44 timestamp
fields, and many of them are not interesting to
index. These rules end up trimming the indexed
fields to just [created_at], and so trailers
are 10x smaller than they would be otherwise.)