all: handle cancellation
When a client HTTP connection disappears
during an expensive query, we should abort
scanning data for that query. We already have
this to the limited extent that queries are
terminated on write errors, but many kinds of
queries (aggregations, etc.) can do tons of
work before writing to the final output connection,
so we cannot rely on that as a cancellation mechanism.

For plan.Transport.Exec, collect parameters into
ExecParams and add plan.ExecParams.Context.

For plan.TableHandle.Open, pass a context.Context
that is canceled when the output pipe disappears.

For plan.Serve, detect cancellation by looking for
a client hangup before the query has completed,
and have plan.Client hang up if its parent context
is canceled.

For tnproto.DirectExec, we return the "error pipe"
as a socket pair so that we can block on the read end
of the pipe inside the tenant process to determine if
the write end has disappeared. (The same applies to
the error pipe produced by tenant.Manager.Do.)

For dcache.Cache.MultiTable, add a context argument
so that we can stop fetching segments if the query
has been canceled.

With all that in place, a client abandoning an HTTP connection
should lead to the entire query pipeline being torn down.