dcache: increase fill parallelism (#1192)
Empirically, the optimal parallelism for cache misses (fills)
is higher than for cache hits; we can only saturate the link
speed with parallelism about equal to 1.5x the number of logical CPUs.

Have cache workers try to spawn new goroutines (up to (GOMAXPROCS+1)/2)
for cache fills, and fall back to blocking when we have reached
capacity.