vm/sort: work around bug in sort.ByColumn
Use ktop sorting wherever possible, since it
supports multiple symbol tables and it is the
most optimized implementation (prefiltering, etc.)
We know LIMIT+OFFSET won't exceed pir.LargeSize (10000)
if we are using this sorting strategy.

For the remaining cases (sorting the output of a GROUP BY
passed through an intermediate projection) we still need
the non-ktop sort path, but in order to work around some
concurrency bugs, simply fall back to stdlib sorting for now.