vm: add support for symbols
Add support for symbols (interned strings)
as values in the bytecode VM.

In order to keep all the existing infrastructure
working more-or-less unchanged, we treat symbols
identically to strings. There are four situations
that need consideration:

 - Hashing
 - Equality comparison (equalv, etc.)
 - Projection
 - String processing

In each of these cases (with the exception of projection),
we insert an "unsymbolize" opcode to re-direct any boxed
values that are symbols to point into the symbol table instead.
(There are only a handful of ops that can produce symbol references,
so we don't have to insert the shim that often.)
For projection, we just project the symbol value as-is,
except for in sorting (because we'd like to treat it as a string).

In order to get good coverage of this behavior,
TestQueries now randomly interns strings in the input.

There is also a new ion.Datum called ion.Interned
which is the stringified representation of an interned string.