ion: Chunker: compress inputs
Compress data insides ion.Chunker.Commit
when the input is found to have repeated strings
that can be interned.

Right now the algorithm is to treat the entirety
of the current block as the "window" for searching
for repeated strings and compressing the window
each time it hits the alignment limit.