ion: harden ReadDatum against adversarial input
We don't yet have an ion API that is designed to
be safe against adversarial input. Since ReadDatum
is the most generic of the ReadXXX methods, add a fuzz
test and fix a bunch of crashes related to broken inputs.
(These aren't currently security issues because we never
call ReadDatum on untrusted data, although we may decide
to do that at some point in the future.)

Also, de-optimize resymbolize() to just a datum decode+encode,
since the optimized path can no longer assume symbols are
produced strictly in-order (this was causing the TestSyntheticRanges
test to fail occasionally).