tenant: output query_error annotation on error
Have the tenant output writer produce an annotation
containing runtime query error information when we
encounter a synchronous error.
The returned annotation looks like

  query_error::{error_message: "..."}

We can't produce this error absolutely 100% reliably
(when the tenant crashes, for example), but it should
be enough to produce an error message produced during
bytecode compilation or subquery substitution.

Also, fix a bug in ion.ReadAnnotation