fixed minor issue with non-ascii ci str compare, added testing of string compare. (#1266)
The bug that is fixed: sStrCmpEqCi assumes a normalized needle that it will compare to data. bug: also non-ASCII values were normalized, thus, when data contains a 'capital omega' and the needle contains the same capital omega, the latter would be normalized to non-capital omega and the compare fails while it should have succeeded.