vm: add more EndSegmentWriter impls
Add more implementations of EndSegmentWriter
and have vm.rowSplitter call them by walking
the list of physical operator implementations.

For alignedWriter, drop the scratch buffer if
it doesn't contain any content.

For bytecode.scratch, save the reserved part
of the scratch in a temporary buffer and then
free it, taking care to restore it the next
time symbolize is called.