vm: simplify string compare test
Move the string comparison bytecode test
into a table-driven test. Also, use groups
of lanes for the unit tests so that the test
runs quickly enough that it can be enabled by default.

The UTF8 test is disabled for now because enabling
multiple lanes of execution reveals some bugs that
need further investagation.