pir: implement more COUNT(DISTINCT ...) cases
Add support for COUNT(DISTINCT ...) as an independent
aggregate function (beyond the limited cases currently
supported).

When we have COUNT(DISTINCT x) with a GROUP BY, we
can create a new sub-query that adds x to the grouping
condition and then re-group the result by x to produce
a table of results we can use with HASH_LOOKUP().
(This is exactly the same technique we use for window
functions; DISTINCT x effectively becomes PARTITION BY x.)

Additionally, ensure that expr.QuoteID quotes identifiers
that contain operator characters and fix up a handful of
test cases that assumed '-' didn't need to be quoted.