zion: "columnar" ion compression
Add a package called "zion"
(for "zipped ion") that performs sort-of-columnar
compression on arbitrary ion data.

The zion format encoded sequences as structures
as multiple compressed data streams. One stream, called the "shape,"
encodes the layout of the structure as sequences of bucket references.
The following streams are zstd-compressed buckets of ion data.

The decoding process uses a user-provided list of paths
to determine which buckets need to be decompressed in advance,
and then walks the "shape" and composes structures with fields
that are members of decompressed buckets.

Empirically, we can "decompress" data up to 6x faster (for 1 field/bucket)
than a naive encoding, although a complete decoding of all fields
ends up running about 2x slower. For common cases (three fields) the
decompression process is still about 2x faster.

This patch also implements the "zion" algorithm for blockfmt.Decoder
so that we can start doing local performance experiments.