vm: relax sorting restrictions
When symbol table changed, check if a new symtab
does not introduce mismatched symbols. When the new
symtab does not change any symbol id, or just add
new symbols, we can carry on sorting.

Only when the new symtab alters the ID of an existing
symbol, we abort sorting procedure with an error.