db: avoid HEAD-ing S3 objects
Use s3.NewFile to construct s3 file handles
when there is enough information presented
to the queue runner to construct the handle
without doing any HEAD operations.

One consequence of not performing a HEAD
up front is that we have to be careful to
handle fs.ErrNotExist and precondition violations
in the conversion code itself, as those will show
up on the first Read call.

Additionally, optimize db.QueueRunner.updateDefs
not to use walkGlob, as it would otherwise list
packfiles we don't want to handle.