ion/blockfmt: filetree: load+store in parallel
Introcude blockfmt.FileTree.Prefetch for performing
the loads associated with a series of Append operations
in parallel, and additionally perform the re-serialization
of the file tree in parallel. These operations each correspond
to one S3 GET or PUT, respectively, so doing a whole bunch
of them serially is pretty slow.