zion: pick seeds via trial histograms
In zion.Encoder.Encode, pick a seed to use
for hashing symbols based on the histogram
of top-level input field sizes.

Since we compute a 64-bit hash function for
mapping symbols to buckets, we can treat the
64-bit hash as sixteen independent trial hashes,
and the "seed" can simply be the nibble position
that produces the best distribution of data.

(We compute the "best" distribution simply
by picking the distribution with the minimum
variance for lack of a better algorithm.)

Also, reserve the remaining 28 bits of "seed"
in the format preamble for other (future) purposes.