ion/blockfmt: filetree: make aliasing explicit
Make the aliasing of level.contents explicit and
encode terminal entries as offsets into the aliased
memory. This reduces the number of Go pointers in
the tree overall (which should reduce memory and GC pressure),
and it makes the memory aliasing explicit, which should
hopefully make it less error-prone.
When leaf nodes are split, we copy only the requisite
terminal entries into the left and right leaves and
discard the originally-aliased memory.