bugfix opContainsPrefixUTF8Ci and opContainsSuffixUTF8Ci (#1527)
bugfixes to opContainsPrefixUTF8Ci and opContainsSuffixUTF8Ci. I realized that the bugs were due to unnecessary complex code that implemented non-essential "fast lane path", that only runs when data and needle both contains only ASCII. The memory layout of the needle is updated such that faster and more logical code is possible. The bug fixes were "free" in this new code.

Note: the code for bcCmpStrEqUTF8Ci is identical to opContainsPrefixUTF8Ci (except for the last line), thus that code is also changed to use the new memory layout of NeedleExt.