plan/pir: optimize ORDER BY positioning
We can abuse the fact that everything following
an aggregation is single-threaded to move ORDER BY
nodes across other steps that don't permute the
input row order (projection and filtering).
Additionally, we can generally shuffle around
LIMIT and BIND steps to guarantee that the ORDER BY
is executed either as part of the hash aggregation itself
or as a ktop ordering rather than an unlimited ordering.