Unified the behavior of a floating-point comparison (#1545)
This commit changes the remaining comparisons (value vs. float and float vs. float)
to behave the same as value vs. value. The reason is that at the moment this was
not symmetric and the results could vary just depending on how we emit opcodes.
Now `NaN == NaN` and `finite_value < NaN` in every scenario.