blockfmt: filetree: set isInner on Append
We set Index.Inputs.root.isInner when index
objects are decoded, but isInner won't be set
if the Index is initialized using a struct literal
(see db/sync.go, db/scan.go, etc.). In that case,
we should make sure that the first Append operation
also sets isInner so that inserting more than 5k
operations before the first roundtrip to S3 doesn't
cause a panic when we try to split the root level.