vm, pir: re-work unnesting
The current unnesting implementation contributes
to a lot of the complexity in plan/pir despite it
not being particulary well-tested.

Re-work unnesting from the ground up so that it
behaves much more straightforwardly and with less
query planner metadata.

The difficulty implementing pir.IterValue is in
introducing a new binding without doing the heavyweight
work of splicing that binding into the current rows.
Instead, we use an "auxilliary binding" in the VM that
is used to pass around values that have been bound without
them being materialized into the record bytes themselves.
(These values are spliced into actual records when they
reach either a projection step or vm.Rematerializer.)

VM ops now need to take *vm.auxbindings into account
when compiling bytecode to see if a binding comes from
*rowParams.auxbound rather than the structures directly.
This data is available from within the bytecode structure
and is accessible using a single instruction.
In order to track the current position within the auxilliary
binding slices, the bytecode "return" instruction now increments
bytecode.auxpos by the number of input lanes, and each vm operator
resets that offset to 0 with bytecode.prepare() in each call
to writeRows().