vm, expr: implement TYPE_BIT
Implement TYPE_BIT, which returns a power-of-2 integer
corresponding to the JSON "type" of the result of an expression.

The purpose of TYPE_BIT is to make it easy and efficient
to aggregate type descriptors using the BIT_OR aggregate function.

Combined with UNPIVOT, you can now ask for a summary of
the types of all the fields of each structure:

  SELECT key, BIT_OR(TYPE_BIT(val))
  FROM UNPIVOT input AS val AT key
  GROUP BY key