plan: merge inputs with unequal predicates
Merge plan inputs under some circumstances when
their predicates are not exactly identical.

In particular, we can safely eliminate parts
of a conjunction from plan.Hints if it doesn't
contain a timestamp comparison, as that's the
only "interesting" hint we can provide in a
predicate right now. See the tests for more details.