vm: add simplification rules
Add rules-based SSA optimizations.

SSA opcodes are either matched literally
or via regexes that match one or more opcodes
(in which case the rule is duplicated for each
matching opcode). Values in non-immediate positions
are matched via unification; values in the immediate
position are matched with == if they are Go string
or number literals.

The code generator consults the ssainfo table
to determine if the argument counts are reasonable,
to line up immediates with their appropriate types,
etc.

The falseprop optimization pass is mostly removed
and replaced with some automatic optimization inside
rewrite() for ops with conjunctive masking semantics.