snellerd: use cgroups cpu controller for tenants
Use the cgroup2 "cpu" controller to encourage
the kernel to split time as fairly as possible
between the different tenant processes.

Mount a "delegated hierarchy" at /sys/fs/cgroup/sneller
and then put each tenant in a sub-hierarchy within it.
The manager has to live in /sys/fs/cgroup/sneller/manager
so that it has a common ancestor that has been delegated
to the unprivileged 'sneller' uid.

Some of the cgroups logic has been pulled out into
a package so that we can re-use it for the ingest
privsep as well.

The snellerd and tenant unit-tests exercise this logic
if you set the CGROOT environment variable to a cgroupv2
root that has been delegated to the uid of the user running
"go test."