tenant: move correct pid to tenant cgroup
The bwrap(1) process that launches the tenant ends
up forking before launching the child so that it
can act as pid 1 in the new pid namespace (and it
does this before blocking on --block-fd).
Consequently, we cannot change the cgroup of the
actual tenant process by using the pid returned
from the initial fork from Go; instead, we have
to use --info-fd to get the pid of the child
and move that instead. (The child is the process
that waits on --block-fd, so there is no race against
the child to move it into the cgroup.)

Tenant tests are updated to double-check that the
tenant lands in the expected place.