db: fix destructive index ops
When merging new data with the latest contents
of blockfmt.Index.Inline via popPrepend(), we
destructively edit blockfmt.Index.Inline, which
means that if we later decide we *don't* need
to do anything with the popped object, the data
is lost.

Change the call sites for popPrepend() so that
we don't make changes to the index until we're
certain we want to, and updating testing to catch
this case.

Also, there's some additional test code for the
objtree since I wrote them in search of this bug.