db: append: invalidate on scan
Invalidate cached index objects if we have
to do a background scan. Scanning is already
a batched operation, so we aren't particularly
interested in whatever performance benefit might
come from caching index between scans, and its
simpler not to have to worry about it.

In practice doing this fixes a condition where
flushScanDone() would write out a new index with
Scanning=false, but the cached copy wouldn't be
updated.