Regex update (#1685)
A update for the regex engine is ready; it features:

A different way of representing end-of-line anchor $ in the NFA/DFA. This allows us to handle regexes that previously we were not supported: e.g. regex a$|b was not but is now supported. A side effect is that the DFA states can be encoded with one bit less, which allows us to handle more complex regexes with faster implementations. However, there still exist trouble makers, e.g. regex ^.*a$|b has issues, the begin-of-line anchor ^ may need a similar treatment as the $.

Unicode wildcards can be handled with ‘tiny’ implementations. Previously, regexes with wildcards .* could only be handled with the ‘large’ implementation. This means that a very large group of regexes can now be handled with the ‘tiny’ implementation, which gives two times faster handling.