tenant/dcache, expr/blob: explicitly mark ephemeral files
The very last file in blockfmt.Index.Inline is updated
"in-place" until it reaches a sufficiently large size,
which means that blobs corresponding to byte-ranges
within that file tend not to be "live" for very long.

Wire up some flags so that cache entries corresponding
to "ephemeral" blobs are preferentially evicted once
they have been dead for a couple seconds.