debug, *: bind debug sockets in children
Have privsep children create a unix socket
that can be used to capture profiling information.

Since we can't have root pre-prepare a listening
socket for each of the children in advance, we can
still sort of emulate a root-owned socket by having
the children insist that the remote end of the connection
has uid 0.

(This still opens up the opportunity for a rogue process
with the right credentials to unlink these debug sockets
or listen on them, but that seems like a mostly-benign threat.)