vm: fix compression of auxvals
The compress_auxvals block in the DISTINCT
and FILTER implementations shouldn't perform
aligned loads, and shouldn't bother loading lanes
that could cross a page boundary if the number
of input lanes is very small.
Use the original input lanes from the VM (k7)
to mask the loads for compression.

Fixes #1721