db: queue: log total queue delay time on inserts
For each table update, log the elapsed time between
the the earliest/latest inserted objects' modtime and
the current time after the index has been updated.

Specifically for S3, the delay stats will encompass
end-to-end latency including the S3 notification delivery
delay, the time it takes us to poll the SQS queue,
and the time it takes to insert the objects into the index.