vm: move scratch info to bytecode ops
Move info about scratch space use to bytecode.go.
Scratch space isn't really a concept we know about
in the SSA code; it's a property of the bytecode ops,
so I think it makes more sense to track it there and
just compute the usage as the program is assembled.