expr: emit MEMBER function only if IN is long enough
If IN contains a short list of constant values, prefer
a series of explicit comparisons.

Comparison of benchmarks before and after introducing
this change.

benchmark                                             old MB/s     new MB/s     speedup
BenchmarkTestQueries/0002-in-in-2-const-ints-8        8564.20      16701.47     1.95x
BenchmarkTestQueries/0002-in-in-3-const-ints-8        8332.01      15637.95     1.88x
BenchmarkTestQueries/0002-in-in-4-const-ints-8        8499.70      15081.69     1.77x
BenchmarkTestQueries/0002-in-in-5-const-ints-8        8448.88      14550.42     1.72x
BenchmarkTestQueries/0002-in-in-6-const-ints-8        8383.69      14175.13     1.69x
BenchmarkTestQueries/0002-in-in-7-const-ints-8        8309.08      13553.17     1.63x
BenchmarkTestQueries/0002-in-in-8-const-ints-8        8437.27      13253.49     1.57x
BenchmarkTestQueries/0002-in-in-9-const-ints-8        7943.94      12793.48     1.61x

BenchmarkTestQueries/0002-in-in-2-const-strings-8     8720.51      15662.94     1.80x
BenchmarkTestQueries/0002-in-in-3-const-strings-8     8490.04      13323.05     1.57x
BenchmarkTestQueries/0002-in-in-4-const-strings-8     8288.61      11881.96     1.43x
BenchmarkTestQueries/0002-in-in-5-const-strings-8     7954.90      10646.74     1.34x
BenchmarkTestQueries/0002-in-in-6-const-strings-8     7744.06      9632.01      1.24x
BenchmarkTestQueries/0002-in-in-7-const-strings-8     7565.19      9017.93      1.19x
BenchmarkTestQueries/0002-in-in-8-const-strings-8     7535.19      8272.96      1.10x
BenchmarkTestQueries/0002-in-in-9-const-strings-8     6842.62      7681.85      1.12x