blockfmt: implement sublinear-time filtering
One long-outstanding issue with the current filter
implementation is that it does not take advantage
of the fact that ranges are ordered, so we can
determine sub-ranges that match a predicate in
logarithmic time rather than linear time.
(Currently, we test that each block sits within
a range that matches the predicate, but we're effectively
repeating a bunch of work at each block.)

Add blockfmt.Filter, which holds compiled predicate
expressions so that sub-ranges within a sparse index
can be iterated efficiently.