db: revert root definition change (#1880)
This reverts a few PRs (#1811, #1834, #1874, #1876) that added
functionality for per-database definition.json files. This retains
changes made to the pattern matching code in the db package.

The reason for this reversion is because per discussion in #1734 we want
to keep support for per table definition.json files, obviate the need to
table name conflict detection, and keep the process of deleting a table
easy.