blockfmt: filter: fix DNF violations
Since blockfmt.Filter implicitly stores the non-empty
intervals to scan in "disjunctive normal form"
(i.e. we want to scan the union of intervals), we need
to be careful when compiling expressions that are not
in DNF. In particular, we need to make sure that negation
(which implicitly produces an AND of ORs) works when applied
to an OR. Otherwise we'd incorrectly union both sides of the
negated OR rather than producing an intersection.

Also, change the structure of filter evaluation to be
continuation-based. For each inner disjunction, the rest
of the filter program is invoked with the latest non-empty
range as the top-of-stack value. So, for (A OR B) AND C,
we implicitly evaluate (A AND C) OR (B AND C) instead, which
keeps things in DNF.

New testcases cover the relevant edge-cases.